requirements.txt
flask
sentence-transformers
faiss-cpu
numpy
requests
lxml
langchain
langchain_ollama
